requirements.txt
asgiref>=3.3.4
Django>=3.2.2
django-bootstrap3>=15.0.0
django-braces>=1.14.0
django-extra-views>=0.13.0
django-filter>=2.4.0
django-query-parameters>=0.2.3
django-sorting-bootstrap>=2.6.2
mysql-connector-python>=8.0.24
mysqlclient>=2.0.3
pytz>=2021.1
six>=1.16.0
sqlparse>=0.4.1
